+++ src/vmath/convert.h
@@ -37,7 +37,7 @@ template <typename T>
 void castString(const vector<string> &s, vector<T> &res) {
     res.resize(s.size());
     T x;
-    for (int i = 0; i < s.size(); ++i) {
+    for (size_t i = 0; i < s.size(); ++i) {
         istringstream ss(s[i]);
         if (!(ss >> x))
             res[i] = -99;
@@ -59,12 +59,12 @@ void transpose(const vector< vector<T> > &data, vector< vector<T> > &cols) {
     int x = data.size();
     int y = data[0].size();
     cols.resize(y);
-    for (int i = 0; i < y; ++i) {
+    for (size_t i = 0; i < y; ++i) {
         cols[i].resize(x);
     }
 
-    for (int i = 0; i < x; i++) {
-        for (int j = 0; j < y; j++) {
+    for (size_t i = 0; i < x; i++) {
+        for (size_t j = 0; j < y; j++) {
             cols[j][i] = data[i][j];
         }
     }
