+++ src/vmath/interp.h
@@ -63,7 +63,7 @@ T interp(T org_x, const vector<T> &temp_x, const vector<T> &temp_y, T low = 0, T
      *TODO - Choose a better search algorithm
      */
     if (sizeTry != temp_x.size()) {
-        for (int i = 0 ; i < (temp_x.size() - 1) ; i++) {
+        for (size_t i = 0 ; i < (temp_x.size() - 1) ; i++) {
             if (temp_x[i] <= org_x && temp_x[i+1] >= org_x) {
                 m = (temp_y[i+1] - temp_y[i])/(temp_x[i+1] - temp_x[i]);
                 c = temp_y[i] - m*temp_x[i];
@@ -81,7 +81,7 @@ template <typename T>
 vector<T> interp(const vector<T> &org_x, const vector<T> &temp_x, const vector<T> &temp_y, T low = 0, T up = 0) {
     vector<T> org_y(org_x.size(), low);
 
-    for (int i = 0; i < org_x.size(); ++i) {
+    for (size_t i = 0; i < org_x.size(); ++i) {
         org_y[i] = interp<T>(org_x[i], temp_x, temp_y, low, up);
     }
